Vuforia Plane Detection/Assets added
Added a few assets and got the plane detection working for the portal.

Don't have stencil masks enabled on floor/roof so we can see where to place objects.

Issue with the room not following the portal (maybe the portal has to be in the center of the room? not sure)